♻️ AppHost | Split into per-host files for readability
AppHost.cs is now a thin orchestrator; each resource and the secrets only it
consumes move into a dedicated file:
- Hosts/CommonInfra.cs  — SQL Server + databases + Azurite (returns a refs record)
- Hosts/WebApiHost.cs   — WebAPI project + its Firebase/SendGrid/SMTP/auth secrets
- Hosts/AdminUiHost.cs  — AdminUI project
- Hosts/DevCommands.cs  — dashboard command buttons + the process helper

Pure refactor — same resource names, env keys, and commands; no behaviour change.
